增加对埋点功能的可扩展性，可以支持第三方埋点（例如：CAT）
