+++ discovery-plugin-strategy-starter-gateway/src/main/java/com/nepxion/discovery/plugin/strategy/gateway/configuration/GatewayStrategyAutoConfiguration.java
@@ -19,7 +19,6 @@ import org.springframework.context.annotation.Configuration;
 import com.nepxion.discovery.plugin.strategy.adapter.DiscoveryEnabledAdapter;
 import com.nepxion.discovery.plugin.strategy.constant.StrategyConstant;
 import com.nepxion.discovery.plugin.strategy.gateway.adapter.DefaultDiscoveryEnabledAdapter;
-import com.nepxion.discovery.plugin.strategy.gateway.context.GatewayStrategyContextHolder;
 import com.nepxion.discovery.plugin.strategy.gateway.filter.GatewayStrategyFilter;
 
 @Configuration
@@ -29,11 +28,6 @@ public class GatewayStrategyAutoConfiguration {
     @Bean
     public GatewayStrategyFilter gatewayStrategyFilter() {
         return new GatewayStrategyFilter();
-    }
-
-    @Bean
-    public GatewayStrategyContextHolder gatewayStrategyContextHolder() {
-        return new GatewayStrategyContextHolder();
     }
 
     @Bean
