+++ discovery-plugin-strategy-starter-gateway/src/main/java/com/nepxion/discovery/plugin/strategy/gateway/configuration/GatewayStrategyContextAutoConfiguration.java
@@ -1,0 +1,26 @@
+package com.nepxion.discovery.plugin.strategy.gateway.configuration;
+
+/**
+ * <p>Title: Nepxion Discovery</p>
+ * <p>Description: Nepxion Discovery</p>
+ * <p>Copyright: Copyright (c) 2017-2050</p>
+ * <p>Company: Nepxion</p>
+ * @author Haojun Ren
+ * @version 1.0
+ */
+
+import org.springframework.boot.autoconfigure.AutoConfigureBefore;
+import org.springframework.cloud.netflix.ribbon.RibbonClientConfiguration;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+import com.nepxion.discovery.plugin.strategy.gateway.context.GatewayStrategyContextHolder;
+
+@Configuration
+@AutoConfigureBefore(RibbonClientConfiguration.class)
+public class GatewayStrategyContextAutoConfiguration {
+    @Bean
+    public GatewayStrategyContextHolder gatewayStrategyContextHolder() {
+        return new GatewayStrategyContextHolder();
+    }
+}
