+++ discovery-plugin-strategy-starter-service/src/main/resources/META-INF/spring.factories
@@ -1,3 +1,4 @@
 org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
 com.nepxion.discovery.plugin.strategy.configuration.StrategyAutoConfiguration,\
-com.nepxion.discovery.plugin.strategy.service.configuration.ServiceStrategyAutoConfiguration
+com.nepxion.discovery.plugin.strategy.service.configuration.ServiceStrategyAutoConfiguration,\
+com.nepxion.discovery.plugin.strategy.service.configuration.ServiceStrategyContextAutoConfiguration
