+++ discovery-plugin-strategy/src/main/java/com/nepxion/discovery/plugin/strategy/rule/DiscoveryEnabledBasePredicate.java
@@ -27,6 +27,10 @@ public class DiscoveryEnabledBasePredicate extends AbstractServerPredicate {
     }
 
     protected boolean apply(Server server) {
+        if (discoveryEnabledAdapter == null) {
+            return true;
+        }
+
         Map<String, String> metadata = pluginAdapter.getServerMetadata(server);
 
         return discoveryEnabledAdapter.apply(server, metadata);
